% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Connect to 'adintel' database when the package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library}

\item{pkgname}{Package name

exportPattern tbl|query|execute|src|^lahman_*}
}
\description{
Connect to 'adintel' database when the package is loaded
}
