% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatter.R
\name{table_formatter}
\alias{table_formatter}
\title{Remove Schema Name from dbcooper Functions}
\usage{
table_formatter(x, schema_name = "public")
}
\arguments{
\item{schema_name}{a string.}

\item{string}{a string or a character vector.}
}
\value{
a string or a character vector.
}
\description{
Remove Schema Name from dbcooper Functions
}
\examples{
x <- c("public.table1", "table2")
table_formatter(x)
}
